Extract caching mechanism into module to avoid repetition
