+++ lib/recruiter/cached_github_repository.rb
@@ -1,30 +1,18 @@
 require 'recruiter/github_repository'
 require 'recruiter/redis_cache'
+require 'recruiter/cache_mechanism'
 
 module Recruiter
   class CachedGithubRepository
+    include CacheMechanism
+
     def initialize(repository, caching_method)
       @composite = repository
       @caching_method = caching_method
     end
 
-    def method_missing(name, *args)
-      cache_key = name.to_s
-      cache_key.concat "_#{args.join("_")}" if args.any?
-
-      if !(elements = @caching_method.fetch(cache_key, @composite.full_name)).nil?
-        cached_elements = elements
-      else
-        elements = args.any? ? @composite.public_send(name, *args) : @composite.public_send(name)
-        @caching_method.store(cache_key, elements, @composite.full_name)
-        cached_elements = elements
-      end
-
-      cached_elements
-    end
-
-    def respond_to_missing?(method_name, include_private = false)
-      @composite.respond_to?(method_name) || super
+    def cache_namespace
+      full_name
     end
 
     def full_name
